Add many missing comments and also links to livestreamer documentation
Add the DocumentationWindow form that only contains a WebBrowser object. The page that should be displayed in the window are given as a parameter when the DocumentationWindow is created.

Add links to "Supported players" and "Supported plugins" sections of the livestreamer documentation that is opened with the new DocumentationWindow form.

Add link to the livestreamer installation page that opens in the default browser.